refactor : 도메인 로직과 ui 로직 분리 및 MVC 패턴 적용한 패키지 구성
model/domain - 비즈니스 로직에 필요한 도메인 클래스(자료 구조)
model/service - 비즈니스 로직을 실행하는 부분
view : UI 로직을 담당
controller : model, view 와 상호작용
validator : exception 처리에 필요한 메소드 구현
