+++ src/main/kotlin/lotto/controller/LottoController.kt
@@ -1,0 +1,4 @@
+package lotto.controller
+
+class LottoController {
+}
