+++ src/main/kotlin/lotto/model/domain/Rank.kt
@@ -1,0 +1,4 @@
+package lotto.model.domain
+
+enum class Rank {
+}
