+++ src/main/kotlin/lotto/model/service/Service.kt
@@ -1,0 +1,4 @@
+package lotto.model.service
+
+class Service {
+}
