+++ src/main/kotlin/lotto/validator/Validator.kt
@@ -1,0 +1,4 @@
+package lotto.validator
+
+class Validator {
+}
