+++ src/main/kotlin/lotto/view/Messages.kt
@@ -1,0 +1,4 @@
+package lotto.view
+
+enum class Messages {
+}
